nsfw-analyzer/requirements.txt
tensorflow==2.13.0
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
flask==2.3.3
werkzeug==2.3.7